+++ apps/web/src/components/review/paginated-list.tsx
@@ -4,7 +4,8 @@ import { useSuspenseQuery } from '@tanstack/react-query';
 
 import ReviewItem from '@/components/review/item.tsx';
 import useMessageToShop from '@/hooks/use-message-to-shop.ts';
-import { type ReviewListFilter, type ReviewListSort, useReviewService } from '@/services/review.tsx';
+import { type ReviewListFilter, type ReviewListSort } from '@/services/api-types/review.tsx';
+import { useReviewService } from '@/services/review.tsx';
 import { useConnectedShop } from '@/state/shop.ts';
 
 interface PaginatedListProps {
